+++ idea/testData/quickfix/createFromUsage/createFunction/call/abstract/interfaceNoExplicitReceiver.kt
@@ -1,0 +1,8 @@
+// "Create abstract function 'foo'" "true"
+interface A {
+    fun bar(b: Boolean) {}
+
+    fun test() {
+        bar(<caret>foo(1, "2"))
+    }
+}