+++ idea/testData/quickfix/createFromUsage/createVariable/parameter/inAccessorInTrait.kt
@@ -2,6 +2,7 @@
 // ACTION: Convert to expression body
 // ACTION: Create local variable 'foo'
 // ACTION: Create property 'foo'
+// ACTION: Create abstract property 'foo'
 // ACTION: Rename reference
 // ERROR: Unresolved reference: foo