+++ idea/testData/quickfix/createFromUsage/createVariable/property/abstract/classNoExplicitReceiver.kt
@@ -1,0 +1,8 @@
+// "Create abstract property 'foo'" "true"
+abstract class A {
+    fun bar(b: Boolean) {}
+
+    fun test() {
+        bar(<caret>foo)
+    }
+}